+++ client/src/App.js
@@ -7,31 +7,9 @@ import 'utils/ignore'
 // assets
 import { imageAssets } from 'theme/images'
 import { fontAssets } from 'theme/fonts'
-import { NativeBaseProvider, extendTheme } from 'native-base'
+import { NativeBaseProvider } from 'native-base'
+import { nativebase } from 'theme'
 import Navigator from './navigator'
-
-// Custom NativeBase theme
-const theme = extendTheme({
-  fontConfig: {
-    GT_Haptik: {
-      regular: {
-        normal: 'GT_Haptik_regular',
-        italic: 'GT_Haptik_regular_italic',
-        rotalic: 'GT_Haptik_rotalic',
-      },
-      bold: {
-        normal: 'GT_Haptik_bold',
-        italic: 'GT_Haptik_bold_italic',
-        rotalic: 'GT_Haptik_rotalic',
-      },
-    },
-  },
-  fonts: {
-    heading: 'GT_Haptik',
-    body: 'GT_Haptik',
-    mono: 'GT_Haptik',
-  },
-})
 
 const App = () => {
   const [didLoad, setDidLoad] = useState(false)
@@ -48,12 +26,12 @@ const App = () => {
 
   return didLoad ? (
     <Provider store={store}>
-      <NativeBaseProvider theme={theme}>
+      <NativeBaseProvider theme={nativebase}>
         <Navigator />
       </NativeBaseProvider>
     </Provider>
   ) : (
-    <NativeBaseProvider theme={theme}>
+    <NativeBaseProvider theme={nativebase}>
       <View />
     </NativeBaseProvider>
   )
