+++ client/src/theme/index.js
@@ -1,9 +1,12 @@
 import colors from './colors'
 import fonts from './fonts'
 import images from './images'
+import nativebase from './nativebase'
 
 /*
     The theme is based on the 7000 Languages Brand Identity Guide.
 */
 
-export { colors, fonts, images }
+export {
+  colors, fonts, images, nativebase,
+}
